refactor: move "list" from `internal/pkg/` to `cmd/devstream/`
